+++ CHANGELOG.rst
@@ -1,5 +1,7 @@
 Ver. 0.5.3 - Development
 ==============================
+* Several fixes in the test application as well as overall refactoring.
+* Tasks throw errors back to the coroutine via the generator ``throw`` method.
 * 50% Faster :class:`pulsar.Deferred` initialisation.
 * Added :meth:`pulsar.Deferred.then` method for adding a deferred to a
   deferred's callbacks without affecting the result.